Refactor test infrastructure and remove BOS dependencies
- Renamed bos_test_gateway to test_servers for clarity
- Removed BOS-specific files (404.html, main.js, bos_test_gateway_server.js)
- Created programmatic test_server.js using httpServer.createServer like original
- Moved test server to test_servers/server/ directory
- Cleaned up package.json scripts removing BOS-related commands
- All E2E tests pass with the new simplified test infrastructure
