+++ package.json
@@ -16,12 +16,10 @@
     "generateimportmap": "(cd public_html && jspm link index.html -o index.html)",
     "serve": "http-server --cors -p 8081 public_html/",
     "dist": "rm -Rf dist && rollup -c rollup.config.js",
-    "link:bos_test_gateway": "(cd bos_test_gateway/ && jspm link 404.html -o 404.html)",
-    "serve:bos_test_gateway": "node bos_test_gateway/server/bos_test_gateway_server.js",
     "create:car": "ipfs-car pack dist/index.html --output dist/index.car",
     "upload:car": "NODE_ENV=mainnet node ./node_modules/nearfs/scripts/upload-car.js dist/index.car",
     "serve:dist": "http-server -p 8081 dist/",
-    "serve:test": "node test_server.js",
+    "serve:test": "node test_servers/server/test_server.js",
     "web4contract": "node web4contract/web4contract.wat.js && wat2wasm web4contract/web4contract.wat -o web4contract/web4contract.wasm",
     "web4contract:deploy": "near contract deploy arizportfolio.near use-file web4contract/web4contract.wasm without-init-call network-config mainnet sign-with-keychain send"
   },
